refactor(FakePlayerResident): streamline bot management and configuration handling
